KAN-85: Rename: Separate socket and Http Meeting - 소켓 사용부분이랑 아닌 부분을 폴더 분리 했습니다.
