+++ src/main/java/capstone/relation/meeting/controller/MeetRoomController.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket.meeting.controller;
+package capstone.relation.meeting.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,11 +7,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import capstone.relation.meeting.dto.request.CreateRoomDto;
+import capstone.relation.meeting.dto.response.JoinResponseDto;
+import capstone.relation.meeting.service.MeetRoomService;
 import capstone.relation.user.UserService;
 import capstone.relation.user.domain.User;
-import capstone.relation.websocket.meeting.dto.request.CreateRoomDto;
-import capstone.relation.websocket.meeting.dto.response.JoinResponseDto;
-import capstone.relation.websocket.meeting.service.MeetRoomService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
