+++ src/main/java/capstone/relation/meeting/repository/MeetRoomRepository.java
@@ -1,11 +1,11 @@
-package capstone.relation.websocket.meeting.repository;
+package capstone.relation.meeting.repository;
 
 import java.util.Set;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import capstone.relation.websocket.meeting.domain.MeetRoom;
+import capstone.relation.meeting.domain.MeetRoom;
 
 @Repository
 public interface MeetRoomRepository extends JpaRepository<MeetRoom, Long> {
