+++ src/main/java/capstone/relation/websocket/signaling/controller/SignalingDocumentController.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket.meeting.controller;
+package capstone.relation.websocket.signaling.controller;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import capstone.relation.websocket.meeting.dto.response.MeetingRoomListDto;
-import capstone.relation.websocket.meeting.dto.signaling.IceDto;
-import capstone.relation.websocket.meeting.dto.signaling.SdpMessageDto;
+import capstone.relation.meeting.dto.response.MeetingRoomListDto;
+import capstone.relation.websocket.signaling.dto.IceDto;
+import capstone.relation.websocket.signaling.dto.SdpMessageDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
