+++ src/main/java/capstone/relation/websocket/signaling/dto/SdpDto.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket.meeting.dto.signaling;
+package capstone.relation.websocket.signaling.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
