+++ src/main/java/capstone/relation/websocket/signaling/dto/SignalMessageType.java
@@ -1,7 +1,7 @@
-package capstone.relation.websocket.meeting.signaling;
+package capstone.relation.websocket.signaling.dto;
 
 public enum SignalMessageType {
-	
+
 	OFFER, ANSWER, ICE, JOIN, LEAVE;
 
 	// enum에서 제공하는 toString()을 사용하면 name()과 동일한 값을 반환합니다.
